backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
numpy==1.26.4
torch==2.4.1
openai-whisper==20231117
webrtcvad==2.0.10

google-cloud-translate==3.15.4
google-cloud-texttospeech==2.16.5
tenacity==9.0.0
